RLC Circuit Analysis and Data Display
An RLC circuit was created in order to calculate, gather and display data regarding the resonant frequency of our given circuit. NI Elvis Instruments were used in order to gather the data while Jupyter was used to create graphs of the data and display the information gathered from the analysis.